requirements.txt
# =============================================================================
# Python依赖包列表 - scripts-for-linux项目
# =============================================================================
#
# 此文件列出了项目运行所需的所有Python第三方依赖包
# 使用方法：pip install -r requirements.txt
#
# 项目支持的Python版本：3.8+
# 测试环境：Ubuntu 20.04-24.04, Debian 10-12
# =============================================================================

# 终端颜色支持库
# 用于：scripts/common.py - 提供跨平台的终端颜色支持
# 可选依赖：如果未安装，将回退到ANSI转义序列
colorama>=0.4.4

# HTTP请求库
# 用于：scripts/containers/docker-push.py, scripts/containers/harbor-push.py
# 功能：Docker Hub API调用、Harbor API调用、镜像搜索等
requests>=2.25.1

# YAML解析库
# 用于：scripts/software/modules/template_coordinator.py
# 功能：解析模块配置文件 modules.yaml
PyYAML>=5.4.1

# =============================================================================
# 开发和测试依赖（可选）
# =============================================================================

# 单元测试框架（Python内置，无需安装）
# unittest

# 代码格式化工具（可选）
# black>=22.0.0

# 代码检查工具（可选）
# flake8>=4.0.0

# 类型检查工具（可选）
# mypy>=0.950

# =============================================================================
# 系统依赖说明
# =============================================================================
#
# 以下系统包需要通过apt安装，不在此requirements.txt中：
# - python3 (>= 3.8)
# - python3-pip
# - python3-venv
# - git
# - curl
# - sudo (用于需要管理员权限的脚本)
#
# 安装命令：
# sudo apt update
# sudo apt install -y python3 python3-pip python3-venv git curl
#
# =============================================================================
# 使用说明
# =============================================================================
#
# 1. 创建虚拟环境：
#    python3 -m venv venv
#
# 2. 激活虚拟环境：
#    source venv/bin/activate  # Linux/macOS
#    venv\Scripts\activate     # Windows
#
# 3. 安装依赖：
#    pip install -r requirements.txt
#
# 4. 运行项目：
#    python3 install.py
#
# 5. 退出虚拟环境：
#    deactivate
#
# =============================================================================
